+++ internal/cli/root.go
@@ -81,7 +81,9 @@ func newRootCmd(stdin io.Reader, stdout, stderr io.Writer) *cobra.Command {
 	root.AddCommand(newMonitorCmd(e))
 	root.AddCommand(newProjectCmd(e))
 	root.AddCommand(newPortalCmd(e))
+	root.AddCommand(newLoadBalancerCmd(e))
 	root.AddCommand(newVolumeCmd(e))
+	root.AddCommand(newGlobalLoadBalancerCmd(e))
 
 	return root
 }